Start basic Git styleguide
This is not intended to be the full encompassing guide on Git, and only
a basic guide. I'd like to provide basic suggestions to the team, in order
to make our commits more consistent. Project teams and individuals can use
this as a guide, and are free to make modifications as they see fit.

Signed-off-by: Doug Yun <doug.yun@dockyard.com>